backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
transformers==4.37.0
torch==2.1.2
onnxruntime==1.17.0
easyocr==1.7.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
validators==0.22.0
tldextract==5.1.1
whois==0.9.27
dnspython==2.5.0
